読み方＆アクセント辞書ダイアログの右側パネルを別コンポーネントにする (#2290)
